Refactor to MiniCart
